+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+## 6.1.0-alpha.1
+
 - feat: Add maxAttachmentSize to SentryOptions #887
 - ref: Remove SentryAttachment.isEqual and hash #885
 - ref: Remove SentryScope.isEqual and hash #884